Initial attempt.
Starting from the code in https://github.com/katakonst/go-dns-proxy,
simplify to a single file which can proxy DNS lookups and replace
A records with AAAA records as described in the SaaS Connector.

This works but is easily derailed by a CNAME. For example,
www.zendesk.com is a CNAME to cdn.cloudflare.com which the
SplitDNS config does not cover and ends up being resolved from
the public Internet.

dns-hijinks needs to be a recursive resolver.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>